+++ CHANGES.md
@@ -1,5 +1,9 @@
 # aws-api
 
+## DEV
+
+* add keyword access to :service (with :metadata) and :http-client
+
 ## 0.8.584 / 2022-08-29
 
 * add keyword access to :region, :endoint, and :credentials on an aws client